% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.metrics.R
\name{basic.metrics}
\alias{basic.metrics}
\title{Basic metrics of the neighboring points.}
\usage{
basic.metrics(space, radius = NULL)
}
\arguments{
\item{space}{A \code{matrix} or \code{data.frame} with xzy coordinates in the first three columns of the neighboring points and a four columns with the distance to the target point.}

\item{radius}{A \code{numeric} vector of a length 1 representing the radius of the sphere that was consider. If \code{radius = NULL}, it use the maximun distance of \code{space} (\code{max(space[,4])}). \code{radius = NULL} as default.}
}
\value{
A \code{data.frame} with four metrics: i) number of neighboring points, ii) volume occupied by neighbors, iii) volume of the sphere used, and iv) density of points based on the volumen of the sphere. If \code{radius = NULL}, the volumnen occupied by neighbors and the volumen of the sphere are the same.
}
\description{
Estimate basic metrics of the surrunding neigboring points created based on a target point.
}
\examples{
data("pc_tree")

neig <- neighborhood(pc_tree[50,], pc_tree, method = "sphere", radius = 0.2)

basic.metrics(neig$neighborhood$`1`, radius = neig$parameter)

}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
