% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_structure.R
\name{canopy_structure}
\alias{canopy_structure}
\title{Canopy Structure}
\usage{
canopy_structure(TLS.type, scan, zenith.range, zenith.rings, azimuth.range,
  vertical.resolution, TLS.resolution, TLS.coordinates = NULL,
  TLS.frame = NULL, TLS.angles = NULL, parallel = FALSE,
  cores = NULL)
}
\arguments{
\item{TLS.type}{A \code{character} describing is the TLS used. It most be one of \code{"single"} return, \code{"multiple"} return, or \code{"fixed.angle"} scanner.}

\item{scan}{If \code{TLS.type} is equal to \code{"single"} or \code{"fixed.angle"}, a \code{data.table} with three columns describing *XYZ* coordinates of the discrete return. If
\code{TLS.type} is equal to \code{"multiple"}, a \code{data.table} with four columns describing *XYZ* coordinates and the target count pulses.}

\item{zenith.range}{If \code{TLS.type} is equal to \code{"single"} or \code{"multiple"}, a \code{numeric} vector of length two describing the \code{min} and \code{max} range of the zenith angle to use.
Theoretically, the \code{max} range should be lower than 90 degrees.}

\item{zenith.rings}{If \code{TLS.type} is equal to \code{"single"} or \code{"multiple"}, a \code{numeric} vector of length one describing the number of zenith rings to use between \code{zenith.range}.
This is used to estimate the frecuency of laser shots from the scanner and returns in \code{scan}. If \code{TLS.type = "fixed.angle"}, \code{zenith.rings = 1} be default.}

\item{azimuth.range}{A \code{numeric} vector of length two describing the range of the azimuth angle to use. Theoretically, it should be between 0 and 360 degrees.}

\item{vertical.resolution}{A \code{numeric} vector of length one describing the vertical resolution to use to extract the profiles from *Z*. Low values lead to more variable profiles.
The scale used needs to be in congruence with the scale of \code{scan}.}

\item{TLS.resolution}{If \code{TLS.type} is equal to \code{"single"} or \code{"multiple"}, a \code{numeric} vector of length two describing the horizontal and vertical angle resolution of the scanner.
If \code{TLS.type} is equal to \code{"fixed.angle"}, a \code{numeric} vector of length one describing the horizontal angle resolution.}

\item{TLS.coordinates}{A \code{numeric} vector of length three describing the scanner coordinates within \code{scan}.
If \code{NULL}, it assumes that the coordinates are \code{c(X = 0, Y = 0, Z = 0)}.}

\item{TLS.frame}{If \code{TLS.type} is equal to \code{"single"} or \code{"multiple"}, a \code{numeric} vector of length four describing the \code{min} and \code{max} of the zenith and azimuth angle of the scanner frame.
If \code{TLS.type = "fixed.angle"}, a \code{numeric} vector of length three describing the fixed zenith angle and the \code{min} and \code{max} of the azimuth angle of the scanner frame.
If \code{NULL}, it assumes that a complete hemisphere (\code{c(zenith.min = 0, zenith.max = 90, azimuth.min = 0, azimuth.max = 360)}), or a cone projection (\code{c(zenith = 57.5, azimuth.min = 0, azimuth.max = 360)}) depending on \code{TLS.type}.}

\item{TLS.angles}{A \code{numeric} vector of length three describing the roll (*X*), pitch (*Y*), and yaw (*Z*) angles of the scanner during the scan.
If \code{NULL}, it assumes that the angles are \code{c(roll = 0, pitch = 0, yaw = 0)}.
This needs to be used if \code{TLS.type} is equal to \code{"single"} or \code{"multiple"}, since it assumes that \code{"fixed.angle"} scanner is previously balanced.}

\item{parallel}{Logical, if \code{TRUE} it use parallel processing on the estimation of shots and returns. \code{FALSE} as default.}

\item{cores}{An \code{integer >= 0} describing the number of cores use. This need to be used if \code{parallel = TRUE}.}
}
\value{
For any \code{TLS.type}, it returns a \code{data.table} with the height profiles defined by \code{vertical.resolution}, the gap probability based on the \code{zenith.range} and \code{zenith.rings}, and
the accumulative L(z) profiles based on the closest zenith ring to 57.5 degrees (hinge angle). If \code{TLS.type} is equal to \code{"fixed.angle"}, it returns f(z) or commonly named PAVD based on
on the ratio of the derivative of L(z) and the derivative of height (z). If \code{TLS.type} is equal to \code{"single"} or \code{"multiple"}, it returns the normalised average weighting L/LAI, and PAVD: based
on the L (hinge angle) at the highest height and the ratio between the derivative of L/LAI average weighted and the derivative of z.
}
\description{
Estimates the canopy structure of a discrete returns scan from different TLS.
}
\details{
Since \code{scan} describes discrete returns measured by the TLS, \code{canopy_structre} first simulates the number of shots emited based on Danson et al. (2007). The simulated shots are
created based on the TLS workflow (\code{TLS.resolution, TLS.frame}) assuming that the scanner is perfectly balance. Then these shots are moved and rotated (\code{\link{move_rotate}}) based on the \code{TLS.angles}
roll, pitch, and yaw, and \code{TLS.coordintates} to simulate the positioning of the scanner during the \code{scan}. Moved and rotated simulated-shots of interest and \code{scan} returns are then extracted based on the \code{zenith.range}, \code{zenith.rings}, and \code{vertical.resolution}.
Using the frecuency of shots and returns the probabiliry of gap (Pgap) is estimated. For \code{TLS.type = "multiple"}, the frecuency of returns is estimated using the sum of 1/target count following Lovell et al. (2011).

Using the Pgap estimated per each zenith ring and vertical profile, \code{canopy_structure} then estimates the accumulative L(z) profiles based on the closest
zenith ring to 57.5 (hinge region) and, if \code{TLS.type} is equal to \code{"fixed.angle"}, the f(z) or commonly named PAVD based on the ratio of the
derivative of L(z) and height (z) following Jupp et al. 2009 (Equation 18). If \code{TLS.type} is equal to \code{"single"} or \code{"multiple"}, \code{canopy_structure} also
estimates the normalised average weighted L/LAI, and then PAVD based on the L (hinge angle) at the highest height (LAI) and the ratio between the derivative
of L/LAI (average weighted) and the derivative of z (Jupp et al. 2009; Equation 21).

Jupp et al. 2009 excludes the zero zenith or fist ring to conduct the average weighted L/LAI estimations, \code{canopy_structre} does not excludes this sections since it depents on the regions of interest of the user.
Therefore, user should consider this difference since it may introduce more variability to profile estimations.
}
\examples{

#Using a multiple return file

\dontrun{
data(TLS_scan)
TLS_scan <- TLS_scan[, 1:4] #Select the four columns required

#This will take a while#
canopy_structure(TLS.type = "multiple", scan = TLS_scan, zenith.range = c(50,70), zenith.rings = 4,
                 azimuth.range = c(0, 360), vertical.resolution = 0.25,
                 TLS.resolution = c(0.04, 0.04), TLS.frame = c(30, 130, 0, 360),
                 TLS.angles =  c(0.293, -0.835, -150.159))

#Using a single return file

data(TLS_scan)
TLS_scan <- TLS_scan[Target_index == 1, 1:3] #Subset to first return observations

#This will take a while#
canopy_structure(TLS.type = "single", scan = TLS_scan, zenith.range = c(50,70), zenith.rings = 4,
                 azimuth.range = c(0, 360), vertical.resolution = 0.25,
                 TLS.resolution = c(0.04, 0.04), TLS.frame = c(30, 130, 0, 360),
                 TLS.angles =  c(0.293, -0.835, -150.159))
}

}
\references{
Danson F.M., Hetherington D., Morsdorf F., Koetz B., Allgower B. 2007. Forest canopy gap fraction from terrestrial laser scanning. IEEE Geosci. Remote Sensing Letters 4:157-160. doi: 10.1109/LGRS.2006.887064.

Lovell J.L., Jupp D.L.B., van Gorsel E., Jimenez-Berni J., Hopkinson C., Chasmer L. 2011. Foliage profiles from ground based waveform and discrete point LiDAR. In: SilviLaser 2011, Hobart, Australia, 16–20 October 2011.

Jupp D.L.B., Culvenor D.S., Lovell J.L., Newnham G.J., Strahler A.H., Woodcock C.E. 2009. Estimating forest LAI profiles and structural parameters using a ground-based laser called “Echidna®”. Tree Physiology 29(2): 171-181. doi: 10.1093/treephys/tpn022.
}
\author{
J. Antonio Guzmán Q.
}
