% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartesian_to_polar.R
\name{cartesian_to_polar}
\alias{cartesian_to_polar}
\title{Cartesian to Polar Coordinates}
\usage{
cartesian_to_polar(cartesian, anchor = NULL)
}
\arguments{
\item{cartesian}{A \code{data.table} with three columns describing the *XYZ* coordinates of a point cloud.}

\item{anchor}{A \code{numeric} vector of length three which describe the *XYZ* anchor coordinate for reference to get the polar coordinates.
If \code{NULL}, it assumes that the reference coordinate is \code{c(X = 0, Y = 0, Z = 0)}.}
}
\value{
A \code{data.table} with the zenith and azimuth angles (degrees), and the distance to the anchor coordinate.
}
\description{
Conver from cartesian to polar coordinates.
}
\details{
It assumes that the positive *Z* axis is the reference vector for the zenith angle. Likewise, it assumes that the *Y* axis is the north-south direction (positive to negative) for the azimuth angle.
If a point from \code{cartesian} presents the same *XY* coordinates than \code{anchor}, \code{angles} returns \code{NA}.
}
\examples{

data(pc_tree)
cartesian_to_polar(pc_tree)
anchor <- c(1, 1, 1)
cartesian_to_polar(pc_tree, anchor)

}
\seealso{
\code{\link{polar_to_cartesian}}
}
\author{
J. Antonio Guzmán Q.
}
