% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_angles.R
\name{cloud_angles}
\alias{cloud_angles}
\title{Zenith and Azimuth of a Point Cloud}
\usage{
cloud_angles(cloud, coordinates)
}
\arguments{
\item{cloud}{A \code{data.table} with three columns describing the *XYZ* coordinates of a point cloud.}

\item{coordinates}{A \code{numeric} vector of length three which describe the *XYZ* reference coordinate.
If \code{NULL}, it assumes that the reference coordinate is \code{c(X = 0, Y = 0, Z = 0)}.}
}
\value{
A \code{data.table} with the zenith and azimuth angles per point in degrees.
}
\description{
Estimates the zenith and azimuth angle of a point cloud based on a reference coordinate.
}
\details{
It assumes that the positive *Z* axis is the reference vector for the zenith angle. Likewise, it assumes that the *Y* axis is the north-south direction (positive to negative) for the azimuth angle.
If a point from \code{cloud} presents the same *XY* coordinates than \code{coordinates}, \code{angles} returns \code{NA}.
}
\examples{

data(pc_tree)
coordinates <- c(mean(pc_tree$X), mean(pc_tree$Y), mean(pc_tree$X))
cloud_angles(pc_tree, coordinates)

}
\author{
J. Antonio Guzmán Q.
}
