% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_dimensionality.R
\name{cloud_dimensionality}
\alias{cloud_dimensionality}
\title{Point cloud dimensionality}
\usage{
cloud_dimensionality(data, method, radius, k, parallel)
}
\arguments{
\item{data}{A \code{matrix} or \code{data.frame} with xyz coordinates in the first three columns or an object of class \code{neighborhood}.}

\item{method}{A character string specifying the method to estimated the neighbors. It most be one of \code{"distance"} or \code{"knn"}.}

\item{radius}{An integer of a length 1 representing the number of neighbors to consider. This will be used if \code{method = "distance"}.}

\item{k}{An integer of a length 1 representing the number of neighbors to consider. This will be used if \code{method = "knn"}.}

\item{parallel}{Logical, if \code{TRUE} it use a parallel processing.}
}
\value{
A \code{data.frame} with the estimated parameters
}
\description{
Estimate the dimensionality of the poits of a cloud. It estimate 10 parameters based on Wang et al. 2017.
}
\examples{
data("pc_tree")

###Run from an object of class data.frame or matrix
cloud_dimensionality(pc_tree, method = "distance", radius = 0.2, parallel = FALSE)

###Run from an object of class neighborhood
dist <- neighborhood(pc_tree, method = "distance", radius = 0.2, parallel = FALSE)
cloud_dimensionality(dist, method = "distance", radius = 0.2, parallel = FALSE)

}
\references{
Wang, D., Hollaus, M., & Pfeifer, N. (2017). Feasibility of machine learning methods for separating wood and leaf points from Terrestrial Laser Scanning data. ISPRS Annals of Photogrammetry, Remote Sensing & Spatial Information Sciences, 4.
}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
