% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_metrics.R
\name{cloud_metrics}
\alias{cloud_metrics}
\title{Point cloud metrics}
\usage{
cloud_metrics(cloud, cloud_b = NULL, basic = TRUE,
  distribution = TRUE, dimensionality = TRUE, method = NULL,
  radius = NULL, k, n_replicates = NULL, parallel = FALSE, cores)
}
\arguments{
\item{cloud}{A \code{data.table} with xyz coordinates in the first three columns or an object of class \code{"neighborhood"}.}

\item{cloud_b}{A \code{data.table} with xyz coordinates in the first three columns. If \code{cloud_b} is \code{NULL}, \code{cloud_b == cloud}. \code{NULL} as default.}

\item{basic}{Logical, if \code{TRUE} it estimate basic metrics. \code{basic = TRUE} as default.}

\item{distribution}{Logical, if \code{TRUE} it estimate distribution metrics of points. \code{distribution = TRUE} as default.}

\item{dimensionality}{Logical, if \code{TRUE} it estimate dimensionality metrics. \code{dimensionality = TRUE} as default.}

\item{method}{A character string specifying the method to estimated the neighbors. It most be one of \code{"sphere"} or \code{"knn"}.}

\item{radius}{A \code{numeric} vector of a length 1 representing the radius of the sphere to consider. This need be used if \code{method = "sphere"} and this may used if \code{method = "knn"}.}

\item{k}{An integer of a length 1 representing the number of neighbors to consider. This need be used if \code{method = "knn"}.}

\item{parallel}{Logical, if \code{TRUE} it use a parallel processing.}

\item{cores}{An \code{integer} >= 0 describing the number of cores use. This need to be used if  \code{parallel = TRUE}.}
}
\value{
A \code{data.frame} with the estimated parameters.
}
\description{
Estimate different metrics on the poits of a cloud. It estimate 10 parameters based on Wang et al. 2017.
}
\examples{
data("pc_tree")

###On objects of class neighborhood
##Calculate the neighborhood of 1000 random rows of a point cloud using the sphere method and a radius of 0.2.
cloud.random <- pc_tree[sample(nrow(pc_tree), 1000), ]
dist <- neighborhood(cloud.random, pc_tree, method = "sphere", radius = 0.2)

#Estimate metrics without using parallel.
cloud_metrics(dist, distribution = FALSE)

#Estimate matrics using parallel processing with 4 cores.
cloud_metrics(dist, distribution = FALSE, parallel = TRUE, cores = 4)

###On objects of class data.table.
##Select 1000 random rows of a point cloud.
cloud.random <- pc_tree[sample(nrow(pc_tree), 1000), ]

#Estimate metrics without using parallel by creating spheres of a radius of 0.2.
cloud_metrics(cloud.random, pc_tree, method = "sphere", radius = 0.2)

#Estimate metrics metrics in parallel using 4 cores by creating spheres of a radius of 0.2.
cloud_metrics(cloud.random, pc_tree, method = "sphere", radius = 0.2, parallel = TRUE, cores = 4)

}
\references{
Wang, D., Hollaus, M., & Pfeifer, N. (2017). Feasibility of machine learning methods for separating wood and leaf points from Terrestrial Laser Scanning data. ISPRS Annals of Photogrammetry, Remote Sensing & Spatial Information Sciences, 4.
}
\seealso{
basic.metrics, distribution, dimensionality
}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
