% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality.R
\name{dimensionality}
\alias{dimensionality}
\title{Dimensionality of the neighboring points.}
\usage{
dimensionality(space)
}
\arguments{
\item{space}{a \code{matrix} or \code{data.frame} with xzy coordinate in the first three columns.}
}
\value{
A \code{data.frame} with the estimated parameters
}
\description{
Calculate the dimensionality of a cloud of neighboring points. It estimate 10 parameters based on Wang et al. 2017.
}
\examples{
data("pc_tree")

#' neig <- neighborhood(pc_tree[50,], pc_tree, method = "sphere", radius = 0.2)
dimensionality(neig$niegborhood$`1`)

}
\references{
Wang, D., Hollaus, M., & Pfeifer, N. (2017). Feasibility of machine learning methods for separating wood and leaf points from Terrestrial Laser Scanning data. ISPRS Annals of Photogrammetry, Remote Sensing & Spatial Information Sciences, 4.
}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
