% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_dimensionality.R
\name{dist_dimensionality}
\alias{dist_dimensionality}
\title{Dimensionality of neighboring points based on distance}
\usage{
dist_dimensionality(x, cloud, radius)
}
\arguments{
\item{x}{a matrix or data.frame of a point with xyz coordinates}

\item{cloud}{a matrix or data.frame of a point cloud with xyz coordinates}

\item{radius}{a numeric vector of a length 1 representing the minimum distance to consider neighboring points}
}
\description{
Estimate the dimensionality of neighboring points of a targed point based on a radius
}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
