% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_neighbors.R
\name{dist_neighbors}
\alias{dist_neighbors}
\title{Neighboors based on distance}
\usage{
dist_neighbors(x, cloud, radius)
}
\arguments{
\item{x}{a targed xyz point.}

\item{cloud}{a point cloud.}

\item{radius}{distance to compute a radius from the targed xyz point. It use the same scalar factor than the point cloud.}
}
\value{
A matrix with the xyz coordinates of the neighboring points and a fourth column with their distance
}
\description{
Extract neighboring points of a targed xyz point based on a radius.
}
