% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstructure.R
\name{fstructure}
\alias{fstructure}
\title{Forest Structure}
\usage{
fstructure(scan, zenith.range, zenith.bands, azimuth.range, TLS.resolution,
  TLS.coordinates = NULL, TLS.frame = NULL, TLS.angles = NULL,
  parallel = FALSE, cores = NULL)
}
\arguments{
\item{scan}{A \code{data.table} with three columns with describe *XYZ* coordinates of the discrete return of a TLS scan.}

\item{zenith.range}{A \code{numeric} vector of length two describing the \code{min} and \code{max} range of the zenith angle to use. Theoretically, the \code{max} range should be lower than 90 degrees.}

\item{zenith.bands}{A \code{numeric} vector of length one describing the number of bands to use between the \code{zenith.range} to estimate the frecuency of laser shots and measured points in \code{scan}.}

\item{azimuth.range}{A \code{numeric} vector of length two describing the range of the azimuth angle to use. Theoretically, it should be between 0 and 360 degrees.}

\item{TLS.resolution}{A \code{numeric} vector of length two describing the horizontal and vertical angle resolution.}

\item{TLS.coordinates}{A \code{numeric} vector of length three the scanner coordinates within \code{returns}.
If \code{NULL}, it assumes that the coordinates are \code{c(X = 0, Y = 0, Z = 0)}.}

\item{TLS.frame}{A \code{numeric} vector of length four describing the \code{min} and \code{max} of the zenith and azimuth angle of the scan frame.
If \code{NULL}, it assumes that a complete hemisphere: \code{c(zenith.min = 0, zenith.max = 90, azimuth.min = 0, azimuth.max = 360)}.}

\item{TLS.angles}{A \code{numeric} vector of length three describing the pitch (*X*), roll (*Y*), and yaw (*Z*) of the TLS during the scan.
If \code{NULL}, it assumes that the angles are \code{c(pitch = 0, roll = 0, yaw = 0)}.}

\item{parallel}{Logical, if \code{TRUE} it use parallel processing on the estimation of shots and returns. \code{FALSE} as default.}

\item{cores}{An \code{integer >= 0} describing the number of cores use. This need to be used if \code{parallel = TRUE}.}
}
\value{
A \code{data.table}
}
\description{
Estimates the forest structure of discrete returns from a TLS scan.
}
\details{

}
\examples{
scan <- fread("C:/Users/josea/Documents/Github/tmp/Example.txt", sep = "\\t")
scan <- scan[, 1:3]
scan[, c("X", "Y", "Z") := round(.SD, 5), .SDcols= c("X", "Y", "Z")]
zenith.range <- c(50.0000, 70.0000)
zenith.bands <- 6
azimuth.range <- c(0.0000, 360.0000)
TLS.resolution <- c(0.04, 0.04)
TLS.coordinates <- c(0, 0, 0)
TLS.frame <- c(30.0000, 150.0000, 0, 360.0000)
TLS.angles <- c(-1.135, 2.434, 147.247)
parallel <- TRUE
cores <- 4

a <- fstructure(scan, zenith.range, zenith.bands, azimuth.range, TLS.resolution, TLS.coordinates, TLS.frame, TLS.angles, parallel, cores )

}
\seealso{
\code{\link{}}
}
\author{
J. Antonio Guzmán Q.
}
