% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_metrics.R
\name{knn_metrics}
\alias{knn_metrics}
\title{k-nearest neighbors metrics on a target point}
\usage{
knn_metrics(x, cloud, k, radius = NULL, basic = TRUE,
  distribution = TRUE, dimensionality = TRUE, n_replicates = NULL)
}
\arguments{
\item{x}{A \code{data.table} or \code{data.frame} of a targed point with xyz coordinates.}

\item{cloud}{A \code{data.table} of a point cloud with xyz coordinates to extract the neighboring points.}

\item{k}{An integer of a length 1 representing the number of neighbors to consider.}

\item{radius}{Optional. A \code{numeric} vector of a length 1 representing a priori radius from \code{x} to select the k nearest neighbors. This speed up the calculations when \code{cloud} is too large.}

\item{basic}{Logical, if \code{TRUE} it estimate basic metrics. \code{basic = TRUE} as default.}

\item{distribution}{Logical, if \code{TRUE} it estimate distribution metrics of points. \code{distribution = TRUE} as default.}

\item{dimensionality}{Logical, if \code{TRUE} it estimate dimensionality metrics. \code{dimensionality = TRUE} as default.}

\item{n_replicates}{A positive \code{interger} of a length 1 representing the number of replicates to estimate the expected distance. If \code{n_replicates = NULL}, it uses the number of neighboring points calculated. \code{n_replicates = NULL} as default.}
}
\value{
A \code{data.table} with the xyz coordinates of the target point and the computed metrics by \code{basic.metrics()}, \code{distribution()}, and \code{dimensionality()} functions.
}
\description{
Estimate different metrics on a neighborhood of a targed point using knn.
}
\examples{
data("pc_tree")
knn_metrics(pc_tree[100,], pc_tree, k = 10)

}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
