% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_distance.R
\name{min_distance}
\alias{min_distance}
\title{Minimun distance between points}
\usage{
min_distance(cloud, parallel = FALSE, cores = NULL)
}
\arguments{
\item{cloud}{A \code{data.table} with xyz coordinates in the first three columns.}

\item{parallel}{Logical, if \code{TRUE} it use a parallel processing. \code{FALSE} as default.}

\item{cores}{An \code{integer} >= 0 describing the number of cores use. This need to be used if \code{parallel = TRUE}.}
}
\value{
A \code{numeric} vector describing the minimun distance between points.
}
\description{
Estimate the minimun distance between points in a point cloud.
}
\examples{
data("pc_tree")

###Estimate the minimun distance of a point cloud
min_distance(pc_tree)

###Estimate the minimun distance of a point cloud using parallel processing with 4 cores
min_distance(pc_tree, parallel = TRUE, cores = 4)

}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
