% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_rotate.R
\name{move_rotate}
\alias{move_rotate}
\title{Moves and Rotates a Point Cloud}
\usage{
move_rotate(cloud, move, rotate)
}
\arguments{
\item{cloud}{A \code{data.table} with three columns describing the *XYZ* coordinates of a point cloud.}

\item{move}{A \code{numeric} vector of length three describing the *XYZ* coordinates to move \code{cloud}.}

\item{rotate}{A \code{numeric} vector of length three describing the rotation angles (degrees) for the roll, pitch, and yaw.}
}
\value{
A \code{data.table} with the rotation and move applied to \code{cloud}.
}
\description{
Moves a point cloud based on *XYZ* coordinates, and rotates it based on the roll, pitch, and yaw angles.
}
\details{
\code{move} conducts a substraction between \code{cloud} less \code{move} coordinates. If \code{NULL}, it does not apply \code{move}.
Likewise, \code{rotate} assumes that roll has an effect on the *X* axis, the pitch on the *Y* axis, and the yaw on the *Z* axis. If \code{NULL}, it does not apply \code{rotate}.
}
\examples{

data(pc_tree)
rgl::plot3d(pc_tree)
coordinates <- c(mean(pc_tree$X), mean(pc_tree$Y), mean(pc_tree$Z))
degrees <- c(45, 45, 0)
rgl::plot3d(move_rotate(pc_tree, coordinates, degrees))

}
\author{
J. Antonio Guzmán Q.
}
