% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood.R
\name{neighborhood}
\alias{neighborhood}
\title{Neighboring points}
\usage{
neighborhood(cloud, method, radius, k, parallel, ...)
}
\arguments{
\item{cloud}{A matrix or data.frame with xyz coordinates in the first three columns.}

\item{method}{A character string specifying the method to estimated the neighbors. It most be one of "distance" or "knn".}

\item{radius}{An integer of a length 1 representing the number of neighbors to consider. This will be used if method = "distance".}

\item{k}{An integer of a length 1 representing the number of neighbors to consider. This will be used if method = "knn".}

\item{parallel}{Logical, if TRUE it use a parallel processing.}
}
\description{
Estimate neighboring points based on two methods.
}
\references{
Wang, D., Hollaus, M., & Pfeifer, N. (2017). Feasibility of machine learning methods for separating wood and leaf points from Terrestrial Laser Scanning data. ISPRS Annals of Photogrammetry, Remote Sensing & Spatial Information Sciences, 4.
}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
