% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_tree-data.R
\docType{data}
\name{pc_tree}
\alias{pc_tree}
\title{A Tree Point Cloud}
\format{
A \code{data.table} with three columns, which are:
\describe{
\item{X}{the "X" coordinate}
\item{Y}{the "Y coordinate}
\item{Z}{the "Z" coordinate}
}

A \code{data.table} where the rows represent the points and the three columns represent the *XYZ* coordinates.
}
\usage{
data(pc_tree)
}
\description{
A \code{data.table} from a point cloud of a tree with a spatial point resolution of 0.05 mm.
}
\examples{
data(pc_tree)
head(pc_tree)

}
\references{
Guzman, Sharp, Alencastro, Sanchez-Azofeifa. 2018. To be published.
}
\keyword{datasets}
