% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.voxels.R
\name{plot.voxels}
\alias{plot.voxels}
\title{Plot method for voxels}
\usage{
plot.voxels(voxels, add.points = TRUE, points.size = 0.1,
  add.voxels = TRUE, border = TRUE, fill = TRUE,
  fillcol = "forestgreen", bordercol = "gray20", lwd = 0.5,
  alpha = 0.1, ...)
}
\arguments{
\item{voxels}{A \code{"voxels"} object.}

\item{add.points}{Logical, if \code{TRUE} it add the original points used to perform the voxalization. \code{TRUE} as default.}

\item{points.size}{The points size, a positive number to use if plot \code{add.points = TRUE}.}

\item{add.voxels}{Logical, if \code{TRUE} it add the voxels created. \code{TRUE} as default.}

\item{border}{Logical, if \code{TRUE} it add a line on the borders of each voxel. \code{TRUE} as default.}

\item{fill}{Logical, if \code{TRUE} it fill each voxels with colors. \code{TRUE} as default.}

\item{fillcol}{A \code{character} definting the color to fill the voxels.}

\item{bordercol}{A \code{character} definting the color of the border of each voxels.}

\item{lwd}{The line width, a positive number, defaulting to 0.5.}

\item{alpha}{A positive numeric vector describing the transparency of the voxels to fill. This value most be between 0.0 (fully transparent) .. 1.0 (opaque).}

\item{...}{General arguments passed to rgl.}

\item{voxel.size}{A positive \code{numeric} vector with the size of the voxel. It use the same dimentional scale of the point cloud.}
}
\description{
The \code{plot} method for voxels created on point clouds using \code{voxels()} function.
}
\examples{
data("pc_tree")

###Create voxels of a size of 0.5.
vox <- voxels(pc_tree, voxel.size = 0.5)
plot.voxels(vox)

}
\seealso{
\code{voxels()}
}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
