% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3d.voxels.R
\name{plot3d.voxels}
\alias{plot3d.voxels}
\title{Plot Method for Voxels}
\usage{
\method{plot3d}{voxels}(voxels, add.points = TRUE, add.voxels = TRUE,
  border = TRUE, fill = TRUE, gradientcol = FALSE, points.size = 1,
  pointscol = "black", fillcol = "forestgreen", lwd = 0.5,
  alpha = 0.1, ...)
}
\arguments{
\item{voxels}{Object of class \code{"voxels"} from \code{\link{voxels}}.}

\item{add.points}{Logical, if \code{TRUE} it adds the original points used to perform the voxalization. \code{TRUE} as default.}

\item{add.voxels}{Logical, if \code{TRUE} it adds the voxels created. \code{TRUE} as default.}

\item{border}{Logical, if \code{TRUE} it adds a line on the borders of each voxel. \code{TRUE} as default.}

\item{fill}{Logical, if \code{TRUE} it fills each voxels with colors. \code{TRUE} as default.}

\item{gradientcol}{Logical, if \code{TRUE} it uses a gradient of colors based on the number of points in each voxels. \code{FALSE} as default.}

\item{points.size}{The points size, a positive number to use if plot \code{add.points = TRUE}.}

\item{pointscol}{A \code{character} definting the color of the points to use.}

\item{fillcol}{A \code{character} definting the color to fill the voxels.}

\item{lwd}{The line width, a positive number, defaulting to 0.5.}

\item{alpha}{A positive numeric vector describing the transparency of the voxels to fill. This value most be between 0.0 (fully transparent) .. 1.0 (opaque).}

\item{...}{General arguments passed to \code{\link{rgl}}.}
}
\description{
The \code{plot} method for objects of class \code{"voxels"} created using the \code{\link{voxels}} function.
}
\examples{
data("pc_tree")

###Create voxels of a size of 0.5.
vox <- voxels(pc_tree, voxel.size = 0.5)
plot3d(vox)

}
\seealso{
\code{\link{voxels}}, \code{\link{voxels_counting}}, \code{\link{summary.voxels}}
}
\author{
J. Antonio Guzmán Q.
}
