% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_to_cartesian.R
\name{polar_to_cartesian}
\alias{polar_to_cartesian}
\title{Polar to Cartesian Coordinates}
\usage{
polar_to_cartesian(polar)
}
\arguments{
\item{polar}{A \code{data.table} with three columns describing the zenith, azimuth, and distance of a point.}
}
\value{
A \code{data.table} with three columns describing the *XYZ* cartesian coordinates
}
\description{
Convert from polar to cartesian coordinates.
}
\examples{

#Creates a hemisphere of points each 2 degrees

zenith <- seq(0, 90, 2)
azimuth <- seq(0, 360, 2)
hemi <- CJ(zenith, azimuth)
hemi$distance <- 1
hemicloud <- polar_to_cartesian(hemi)
rgl::plot3d(hemicloud)

}
\seealso{
\code{\link{}}
}
\author{
J. Antonio Guzmán Q.
}
