% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotate a Point Cloud}
\usage{
rotate(cloud, roll = 0, pitch = 0, yaw = 0)
}
\arguments{
\item{cloud}{A \code{data.table} with three columns describing the *XYZ* coordinates of a point cloud.}

\item{roll}{A \code{numeric} vector describing the degrees of rotation angles for roll (*X*).}

\item{pitch}{A \code{numeric} vector describing the degrees of rotation angles for pitch (*Y*).}

\item{yaw}{A \code{numeric} vector describing the degrees of rotation angles for yaw (*Z*). for the roll, pitch, and yaw.}
}
\value{
A \code{data.table} with the rotation applied to \code{cloud}.
}
\description{
Rotate point cloud based on the roll, pitch, and yaw angles.
}
\details{
The *XYZ* coordiantes are transformed to E-N-U coordinates (ENU system, East-North-Up).
}
\examples{

data(pc_tree)
rgl::plot3d(pc_tree)
rgl::plot3d(rotate(pc_tree, roll = 45, pitch = 45, yaw = 0))

}
\author{
J. Antonio Guzmán Q.
}
