% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTLS-package.R
\docType{package}
\name{rTLS-package}
\alias{rTLS}
\alias{rTLS-package}
\title{rTLS: Tools to Process Point Clouds Derived From Terrestrial Laser Scanning}
\description{
rTLS is a package that compiles a set of tools to process and calculate
metrics on point clouds derived from terrestrial
LiDAR (Light Detection and Ranging). Its creation is based on key
aspects of the TLS application in forestry. Currently, the main
routines are based on filtering, neighboring features of points,
voxelization, optimal sphere or voxel size, and the creation of
artificial stands. rTLS is written using data.table and C++ language and in most of the
functions it is posible to use parallel processing to speed-up the routines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Antguz/rTLS}
  \item Report bugs at \url{https://github.com/Antguz/rTLS/issues}
}

}
\author{
\strong{Maintainer}: J. Antonio Guzmán Q. \email{antguz06@gmail.com} (\href{https://orcid.org/0000-0002-0721-148X}{ORCID}) [thesis advisor, copyright holder]

Authors:
\itemize{
  \item J. Antonio Guzmán Q. \email{antguz06@gmail.com} (\href{https://orcid.org/0000-0002-0721-148X}{ORCID}) [thesis advisor, copyright holder]
  \item Ronny Hernandez \email{ronny.hernandezm@gmail.com} (\href{https://orcid.org/0000-0001-6225-7096}{ORCID})
}

Other contributors:
\itemize{
  \item Arturo Sanchez-Azofeifa \email{arturo.sanchez@ualberta.ca} (\href{https://orcid.org/0000-0001-7768-6600}{ORCID}) [degree supervisor, scientific advisor]
}

}
