% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere_dimensionality.R
\name{sphere_dimensionality}
\alias{sphere_dimensionality}
\title{Dimensionality of neighboring points based on a radius of a sphere.}
\usage{
sphere_dimensionality(x, cloud, radius)
}
\arguments{
\item{x}{A \code{matrix} or \code{data.frame} of a point with xyz coordinates.}

\item{cloud}{a \code{matrix} or \code{data.frame} of a point cloud with xyz coordinates.}

\item{radius}{a numeric vector of a length 1 representing the minimum distance to consider neighboring points.}
}
\value{
A  \code{matrix} with the xyz coordinates of the neighboring points and a fourth column with their distance.
}
\description{
Estimate the dimensionality of the neighboring points of a targed point based on a radius.
}
\examples{
data("pc_tree")
dist_dimensionality(pc_tree[100,], pc_tree, radius = 0.2)

}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
