% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere_neighbors.R
\name{sphere_neighbors}
\alias{sphere_neighbors}
\title{Neighboring points in a sphere}
\usage{
sphere_neighbors(x, cloud, radius)
}
\arguments{
\item{x}{A \code{matrix} or \code{data.frame} of the targed point with xyz coordinates.}

\item{cloud}{A \code{matrix} or \code{data.frame} of a point cloud with xyz coordinates to extract the neighboring points.}

\item{radius}{A numeric \code{vector} of a length 1 representing the minimum distance to consider neighboring points.}
}
\value{
A \code{data.frame} with the xyz coordinates of the neighboring points and a fourth column with their distance.
}
\description{
Estimate the neighboring points in a shpere around the targed point.
}
\examples{
data("pc_tree")
sphere_neighbors(pc_tree[100,], pc_tree, radius = 0.5)

}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
