% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_neighborhood.R
\name{sub_neighborhood}
\alias{sub_neighborhood}
\title{Subset neighboring points}
\usage{
sub_neighborhood(x, new_distance)
}
\arguments{
\item{x}{An object of class \code{neighborhood}.}

\item{new_distance}{A numeric vector that swhich specifies the new distance to subset.}
}
\value{
An object of class "neighborhood" with a list per point of the neighboring points.
}
\description{
Subset neighboring points based on a distance or knn
}
\examples{
data("pc_tree")

###Subset neighboring points from 0.2 to 0.1
dis <- neighborhood(pc_tree, method = "distance", radius = 0.2, parallel = FALSE)
sub_neighborhood(dis, new_distance = 0.1)

}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
