% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_metrics.R
\name{tree_metrics}
\alias{tree_metrics}
\title{Tree metrics}
\usage{
tree_metrics(cloud, region.diameter = NULL)
}
\arguments{
\item{cloud}{A \code{data.table} with xyz coordinates in the first three columns.}

\item{region.diameter}{A numeric \code{vector} of length 2 indicating the lower and higher region to subset the point cloud and get the diameter. If \code{cloud_b} with xyz coordinates in the first three columns. If \code{region.diameter = NULL}, it use \code{c(1.25, 1.35)}. \code{NULL} as default.}
}
\value{
A \code{data.table} with the tree hight, crown area, and diameter
}
\description{
Estimate the tree height, crown area, and diameter at breast height
}
\examples{
data("pc_tree")
tree_metrics(pc_tree)

}
\seealso{
cloud_metrics
}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
