% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_structure.R
\name{veg_structure}
\alias{veg_structure}
\title{Vegetation structure}
\usage{
veg_structure(voxels, k = NULL)
}
\arguments{
\item{voxels}{A \code{"voxels"} object.}

\item{k}{A coefficient of extinction. If \code{NULL}, 0.5 is assumed as a proxy to characterize the vegetation properties.}
}
\value{
If \code{obj.voxel = TRUE}, it return an object of class \code{"voxels"} wich contain a list with the points used to create the voxels, the parameter \code{voxel.size}, and the \code{voxels} created. If \code{obj.voxel = FALSE}, it returns a \code{data.table} with the coordinates of the voxels created and the number of points in each voxel.
}
\description{
Estimates the vegetation structure of the vertical profile of a point cloud based using voxel approach.
}
\examples{
data("pc_tree")

###Create voxels of a size of 0.5.
vox <- voxels(pc_tree, voxel.size = 0.5)

###Estimate the vegetation structure
veg_structure(vox)

}
\references{
Bouvier, M., Durrieu, S., Fournier, R. a, & Renaud, J. (2015).  Generalizing predictive models of forest inventory attributes using an area-based approach with airborne las data. Remote Sensing of Environment, 156, 322-334. http://doi.org/10.1016/j.rse.2014.10.004
}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
