% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_metrics.R
\name{cloud_metrics}
\alias{cloud_metrics}
\title{Point cloud metrics}
\usage{
cloud_metrics(cloud, cloud_b = NULL, method, radius = NULL, k,
  basic = TRUE, distribution = TRUE, dimensionality = TRUE,
  n_replicates = NULL, parallel = FALSE, cores)
}
\arguments{
\item{cloud}{A \code{data.table} with *XYZ* coordinates in the first three columns or an object of class \code{"neighborhood"}.}

\item{cloud_b}{A \code{data.table} with *XYZ* coordinates in the first three columns. If \code{cloud_b} is \code{NULL}, \code{cloud_b == cloud}. \code{NULL} as default.}

\item{method}{A character string specifying the method to estimate the neighbors. It most be one of \code{"sphere"} or \code{"knn"}.}

\item{radius}{A \code{numeric} vector of a length 1 representing the radius of the sphere to consider. This needs be used if \code{method = "sphere"} and may used if \code{method = "knn"}.}

\item{k}{An integer of a length 1 representing the number of neighbors to consider. This needs be used if \code{method = "knn"}.}

\item{basic}{Logical, if \code{TRUE} it estimates \code{\link{basic_metrics}}. \code{TRUE} as default.}

\item{distribution}{Logical, if \code{TRUE} it estimates \code{\link{distribution}} metrics of points. \code{TRUE} as default.}

\item{dimensionality}{Logical, if \code{TRUE} it estimates \code{\link{dimensionality}} metrics. \code{TRUE} as default.}

\item{n_replicates}{An \code{interger} of a length 1 representing the number of replicates to estimate the expected distance. This needs to be used if \code{distribution = TRUE}.If \code{NULL}, it uses the number of neighbors per point. \code{NULL} as default.}

\item{parallel}{Logical, if \code{TRUE} it uses parallel processing.}

\item{cores}{An \code{integer} >= 0 describing the number of cores use. This needs to be used if  \code{parallel = TRUE}.}
}
\value{
A \code{data.frame} with the estimated parameters.
}
\description{
Estimate different metrics on the poits of a cloud.
}
\examples{
data("pc_tree")

###On objects of class neighborhood
#Calculate the neighborhood on a sample the sphere method with a 0.2 of radius.
cloud.random <- pc_tree[sample(nrow(pc_tree), 1000), ]
dist <- neighborhood(cloud.random, pc_tree, method = "sphere", radius = 0.2)

#Estimate metrics without using parallel.
cloud_metrics(dist)

\dontrun{
#Estimate matrics using parallel processing with 4 cores.
cloud_metrics(dist, parallel = TRUE, cores = 4)
}

###On objects of class data.table.
##Select 1000 random rows of a point cloud.
cloud.random <- pc_tree[sample(nrow(pc_tree), 1000), ]

#Estimate metrics without using parallel by creating spheres of a radius of 0.2.
cloud_metrics(cloud.random, pc_tree, method = "sphere", radius = 0.2)

\dontrun{
#Estimate metrics metrics in parallel using 4 cores by creating spheres of a radius of 0.2.
cloud_metrics(cloud.random, pc_tree, method = "sphere", radius = 0.2, parallel = TRUE, cores = 4)
}

}
\seealso{
\code{\link{basic_metrics}}, \code{\link{distribution}}, \code{\link{dimensionality}}, \code{\link{neighborhood}}
}
\author{
J. Antonio Guzmán Q. and Ronny Hernandez
}
