% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_metrics.R
\name{cloud_metrics}
\alias{cloud_metrics}
\title{Point cloud metrics}
\usage{
cloud_metrics(cloud, cloud_b = NULL, basic = TRUE,
  distribution = TRUE, dimensionality = TRUE, method = NULL,
  radius = NULL, k, n_replicates = NULL, parallel = FALSE,
  threads = NULL)
}
\arguments{
\item{cloud}{A \code{data.table} with xyz coordinates in the first three columns or an object of class \code{neighborhood}.}

\item{cloud_b}{A \code{data.table} with xyz coordinates in the first three columns. If \code{cloud_b} is \code{NULL}, \code{cloud_b == cloud}. \code{NULL} as default.}

\item{basic}{Logical, if \code{TRUE} it estimate basic metrics. \code{basic = TRUE} as default.}

\item{distribution}{Logical, if \code{TRUE} it estimate distribution metrics of points. \code{distribution = TRUE} as default.}

\item{dimensionality}{Logical, if \code{TRUE} it estimate dimensionality metrics. \code{dimensionality = TRUE} as default.}

\item{method}{A character string specifying the method to estimated the neighbors. It most be one of \code{"sphere"} or \code{"knn"}.}

\item{radius}{A \code{numeric} vector of a length 1 representing the radius of the sphere to consider. This need be used if \code{method = "sphere"} and this may used if \code{method = "knn"}.}

\item{k}{An integer of a length 1 representing the number of neighbors to consider. This need be used if \code{method = "knn"}.}

\item{parallel}{Logical, if \code{TRUE} it use a parallel processing.}

\item{threads}{An \code{integer} >= 0 describing the number of CPU use. This need to be used if  \code{parallel = TRUE}}
}
\value{
A \code{data.frame} with the estimated parameters.
}
\description{
Estimate different metrics on the poits of a cloud. It estimate 10 parameters based on Wang et al. 2017.
}
\examples{
data("pc_tree")

###Run on the first 100 points of a point cloud of class data.table using a sphere method with a radius of 0.2
cloud_metrics(pc_tree[1:100,], pc_tree, method = "sphere", radius = 0.2)

###Run on the first 100 points of a point cloud of class data.table using a knn method with a k = 10.
cloud_metrics(pc_tree[1:100,], pc_tree, method = "knn", k = 10)


###Estimate neighborhood on first 100 points using a sphere method and a radius of 0.2 and then calculate their metrics
dist <- neighborhood(pc_tree[1:100,], pc_tree, method = "sphere", radius = 0.2)
cloud_metrics(dist)

}
\references{
Wang, D., Hollaus, M., & Pfeifer, N. (2017). Feasibility of machine learning methods for separating wood and leaf points from Terrestrial Laser Scanning data. ISPRS Annals of Photogrammetry, Remote Sensing & Spatial Information Sciences, 4.
}
\seealso{
basic.metrics, distribution, dimensionality
}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
