% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_metrics.R
\name{cloud_metrics}
\alias{cloud_metrics}
\title{Point cloud metrics}
\usage{
cloud_metrics(cloud, cloud_b = NULL, method, radius, k,
  parallel = FALSE)
}
\arguments{
\item{cloud}{A \code{matrix} or \code{data.frame} with xyz coordinates in the first three columns or an object of class \code{neighborhood}.}

\item{cloud_b}{A \code{matrix} or \code{data.frame} with xyz coordinates in the first three columns. If \code{cloud_b} is \code{NULL}, \code{cloud_b == cloud}. \code{NULL} as default.}

\item{method}{A character string specifying the method to estimated the neighbors. It most be one of \code{"sphere"} or \code{"knn"}.}

\item{radius}{A \code{numeric} vector of a length 1 representing the radius of the sphere to consider. This will be used if \code{method = "sphere"}.}

\item{k}{An integer of a length 1 representing the number of neighbors to consider. This will be used if \code{method = "knn"}.}

\item{parallel}{Logical, if \code{TRUE} it use a parallel processing.}
}
\value{
A \code{data.frame} with the estimated parameters.
}
\description{
Estimate different metrics on the poits of a cloud. It estimate 10 parameters based on Wang et al. 2017.
}
\examples{
data("pc_tree")

###Run from an object of class data.frame or matrix
cloud_metrics(pc_tree, method = "sphere", radius = 0.2, parallel = FALSE)

###Run from an object of class neighborhood
dist <- neighborhood(pc_tree, method = "sphere", radius = 0.2, parallel = FALSE)
cloud_metrics(dist, method = "sphere", radius = 0.2, parallel = FALSE)

}
\references{
Wang, D., Hollaus, M., & Pfeifer, N. (2017). Feasibility of machine learning methods for separating wood and leaf points from Terrestrial Laser Scanning data. ISPRS Annals of Photogrammetry, Remote Sensing & Spatial Information Sciences, 4.
}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
