% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality.R
\name{dimensionality}
\alias{dimensionality}
\title{Dimensionality of the Neighboring Points.}
\usage{
dimensionality(space)
}
\arguments{
\item{space}{A \code{data.table} with *XYZ* coordinates of the neighboring points in the first three columns.}
}
\value{
A \code{data.table} with the estimated parameters.
}
\description{
Calculate the dimensionality of a cloud of neighboring points. It estimates nine parameters based on Wang et al. 2017.
}
\details{
Since the extraction of the metrics are based on \code{\link{eigen}}, a \code{space} of neighboring points lower than three may return \code{NA} as results.
}
\examples{
data("pc_tree")
neig <- neighborhood(pc_tree[50,], pc_tree, method = "sphere", radius = 0.2)
dimensionality(neig$neighborhood[, c(2:4)])

}
\references{
Wang, D., Hollaus, M., & Pfeifer, N. (2017). Feasibility of machine learning methods for separating wood and leaf points from Terrestrial Laser Scanning data. ISPRS Annals of Photogrammetry. Remote Sensing & Spatial Information Sciences, 4. <doi:10.5194/isprs-annals-IV-2-W4-157-2017>
}
\seealso{
\code{\link{basic_metrics}}, \code{\link{distribution}}, \code{\link{cloud_metrics}}, \code{\link{neighborhood}}
}
\author{
J. Antonio Guzmán Q. and Ronny Hernandez
}
