% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality.R
\name{dimensionality}
\alias{dimensionality}
\title{Dimensionality of neighboring points}
\usage{
dimensionality(space)
}
\arguments{
\item{space}{a matrix or data.frame with xzy coordinate in the first three columns.}
}
\description{
Calculate the dimensionality of a cloud of neighboring points. It estimate 10 parameters based on Wang et al. 2017.
}
