% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution}
\alias{distribution}
\title{Dispersion, agregation, and mean distance of neighboring points.}
\usage{
distribution(space, radius = NULL, n_replicates = NULL)
}
\arguments{
\item{space}{A \code{matrix} or \code{data.frame} with xzy coordinates in the first three columns of the neighboring points and a four columns with the distance to the target point.}

\item{radius}{A \code{numeric} vector of a length 1 representing the radius of the sphere which was considered. If \code{radius = NULL}, it use the maximun distance of \code{space} (\code{max(space[,4])}). \code{radius = NULL} as default.}
}
\value{
A \code{data.frame} with four metrics: i) observed distance, ii) expected distance, iii) points dispersion, iv) aggregation to the target point.
}
\description{
Estimate the dispersion, aggregation, and mean distance of the neighboring points to a target point..
}
\examples{
data("pc_tree")

neig <- neighborhood(pc_tree[50,], pc_tree, method = "sphere", radius = 0.2)

distribution(neig$neighborhood[, c(2:5)], radius = neig$parameter)

}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
