% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{Point cloud filter}
\usage{
filter(cloud, method, k, radius = NULL, nSigma, mnp, parallel = FALSE,
  cores = NULL, plot = TRUE)
}
\arguments{
\item{cloud}{A \code{data.table} with xyz coordinates in the first three columns.}

\item{method}{A character string specifying the method to estimated the neighbors. It most be one of \code{"sor"} for Statistical Outlier Removal or \code{"mnp"} for Minimun number of points.}

\item{k}{An integer of a length 1 representing the number of neighboring points to consider. This need to be used if \code{method = "sor"}.}

\item{radius}{A \code{numeric} vector of a length 1 representing the distance to consider for the extraction of number of points. This need to be used if \code{method = "mnp"} and this could be used if \code{method = "sor"}.}

\item{nSigma}{A \code{numeric} vector of a length 1 representing the standard deviation multiplier.}

\item{parallel}{Logical, if \code{TRUE} it use a parallel processing. \code{FALSE} as default.}

\item{plot}{Logical, if \code{TRUE} returns a plot showing the parameters used for filtering.}
}
\value{
A filtered \code{data.table}.
}
\description{
Perform a different filtering methods on a point cloud.
}
\examples{
data("pc_tree")

###Filter a cloud using the SOR method with a k of 20 and nSigma of 2
filter(pc_tree, method = "sor", k = 20, nSigma = 2, parallel = TRUE, cores = 8)

###Filter a cloud using the MNP method with a radius of 0.5 and mnp of 5.
filter(pc_tree, method = "mnp", radius = 0.5, mnp = 100, parallel = TRUE, cores = 8)

}
\references{
Wang, D., Hollaus, M., & Pfeifer, N. (2017). Feasibility of machine learning methods for separating wood and leaf points from Terrestrial Laser Scanning data. ISPRS Annals of Photogrammetry, Remote Sensing & Spatial Information Sciences, 4.
}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
