% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_dimensionality.R
\name{knn_dimensionality}
\alias{knn_dimensionality}
\title{Dimensionality of neighboring points based on knn}
\usage{
knn_dimensionality(x, cloud, k)
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame} of a point with xyz coordinates}

\item{cloud}{an \code{matrix} or \code{data.frame} of a point cloud with xyz coordinates}

\item{radius}{an integer of a length 1 representing the number of neighbors to consider}
}
\value{
A \code{matrix} with the xyz coordinates of the neighboring points and a fourth column with their distance
}
\description{
Estimate the dimensionality of neighboring points of a targed point based on knn
}
\examples{
data("pc_tree")
knn_dimensionality(pc_tree[100,], pc_tree, k = 10)

}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
