% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_neighbors.R
\name{knn_neighbors}
\alias{knn_neighbors}
\title{Neighboring points based on knn}
\usage{
knn_neighbors(targed, cloud, k, radius = NULL)
}
\arguments{
\item{cloud}{A \code{data.table} of a point cloud with xyz coordinates.}

\item{k}{An integer of a length 1 representing the number of neighbors to consider.}

\item{radius}{Optional. A \code{numeric} vector of a length 1 representing a priori radius from \code{x} to select the k nearest neighbors.   number of neighbors to consider. It speed up the calculations when \code{cloud} is too large.}

\item{x}{A \code{data.frame} or \code{data.table} of a point with xyz coordinates.}
}
\value{
A \code{matrix} with the xyz coordinates of the k neighboring points and a fourth column with their distance.
}
\description{
Estimate the neighboring points of a targed point based on a knn.
}
\examples{
data("pc_tree")
knn_neighbors(pc_tree[100,], pc_tree, k = 10)
knn_neighbors(pc_tree[100,], pc_tree, k = 10, radius = 0.5)

}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
