% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_neighbors.R
\name{knn_neighbors}
\alias{knn_neighbors}
\title{Neighboring points based on knn}
\usage{
knn_neighbors(x, cloud, k)
}
\arguments{
\item{x}{a matrix or data.frame of a point with xyz coordinates}

\item{cloud}{a matrix or data.frame of a point cloud with xyz coordinates}

\item{k}{an integer of a length 1 representing the number of neighbors to consider}
}
\description{
Estimate the neighboring points of a targed point based on a knn
}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
