% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_cloud_metrics.R
\name{m_cloud_metrics}
\alias{m_cloud_metrics}
\title{Point Cloud Metrics at Multiple Scales}
\usage{
m_cloud_metrics(
  cloud,
  cloud_b = NULL,
  basic = TRUE,
  distribution = TRUE,
  dimensionality = TRUE,
  method = NULL,
  radius.range = NULL,
  k.range,
  n_replicates = NULL,
  parallel = FALSE,
  cores = NULL
)
}
\arguments{
\item{cloud}{An object of class \code{"neighborhood"} or \code{data.table} with *XYZ* coordinates in the first three columns.}

\item{cloud_b}{A \code{data.table} with *XYZ* coordinates in the first three columns. If \code{cloud_b = NULL}, \code{cloud_b == cloud}. \code{NULL} as default.}

\item{basic}{Logical, if \code{TRUE} it estimates \code{\link{basic_metrics}}. \code{basic = TRUE} as default.}

\item{distribution}{Logical, if \code{TRUE} it estimates \code{\link{distribution}} metrics of points. \code{distribution = TRUE} as default.}

\item{dimensionality}{Logical, if \code{TRUE} it estimates \code{\link{dimensionality}} metrics. \code{dimensionality = TRUE} as default.}

\item{method}{A \code{character} string specifying the method to estimate neighbors. It most be one of \code{"sphere"} or \code{"knn"}.}

\item{radius.range}{A positive \code{numeric} vector representing the range of spheres radius to consider. This need be used if \code{method = "sphere"}. This may be used if \code{method = "knn"}, but this should to be a \code{numeric} vector of length 1.}

\item{k.range}{An \code{integer} of a length 1 representing the number of neighbors to consider. This need be used if \code{method = "knn"}.}

\item{n_replicates}{An \code{interger} of a length 1 representing the number of replicates to estimate the expected distance. This needs to be used if \code{distribution = TRUE}.If \code{NULL}, it uses the number of neighbors per point. \code{NULL} as default.}

\item{parallel}{Logical, if \code{TRUE} it use a parallel processing.}

\item{cores}{An \code{integer} >= 0 describing the number of cores use. This need to be used if \code{parallel = TRUE}.}
}
\value{
A \code{data.table} with the estimated parameters.
}
\description{
Estimate different metrics on the poits of a cloud using multiple scales.
}
\examples{

data("pc_tree")

###On objects of class neighborhood
##Calculate the neighborhood on a sample using the sphere method with a 0.2 of radius.
cloud.random <- pc_tree[sample(nrow(pc_tree), 100), ]
dist <- neighborhood(cloud.random, pc_tree, method = "sphere", radius = 5)

#Estimate on three different scales scales of spheres (5, 2.5) without parallel
m_cloud_metrics(dist, radius.range = c(5, 2.5), n_replicates = 10, parallel = FALSE)

\dontrun{
#Estimate metrics using three different scales of spheres (5, 2.5) with parallel
m_cloud_metrics(dist, radius.range = c(5, 2.5), n_replicates = 10, parallel = TRUE, cores = 4)
}

}
\seealso{
\code{\link{basic_metrics}}, \code{\link{distribution}}, \code{\link{dimensionality}}, \code{\link{cloud_metrics}}, \code{\link{neighborhood}}
}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
