% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_distance.R
\name{min_distance}
\alias{min_distance}
\title{Minimun Distance Between Points}
\usage{
min_distance(cloud, threads = 1, progress = TRUE)
}
\arguments{
\item{cloud}{A \code{data.table} with *XYZ* coordinates in the first three columns representing a point cloud.}

\item{threads}{An \code{integer} specifying the number of threads to use for parallel processing. Experiment to see what works best for your data on your hardware.}

\item{progress}{Show progress. \code{TRUE} as default.}
}
\value{
A \code{numeric} vector describing the minimun distance between points.
}
\description{
Estimate the minimum distance between points in a point cloud.
}
\examples{
data("pc_tree")

#Estimate the minimun distance of a sample o 100 points
cloud_sample <- pc_tree[sample(nrow(pc_tree), 100), ]
min_distance(cloud_sample)

}
\author{
J. Antonio Guzmán Q.
}
