% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood.R
\name{neighborhood}
\alias{neighborhood}
\title{Neighboring points}
\usage{
neighborhood(cloud, cloud_b = NULL, method, radius, k, parallel = NULL)
}
\arguments{
\item{cloud}{A \code{matrix} or \code{data.frame} with xyz coordinates in the first three columns.}

\item{cloud_b}{A \code{matrix} or \code{data.frame} with xyz coordinates in the first three columns. If \code{cloud_b == NULL}, \code{cloud_b == cloud}. \code{NULL} as default.}

\item{method}{A character string specifying the method to estimated the neighbors. It most be one of  \code{"sphere"} or  \code{"knn"}.}

\item{radius}{An integer of a length 1 representing the number of neighbors to consider. This will be used if  \code{method = "sphere"}.}

\item{k}{An integer of a length 1 representing the number of neighbors to consider. This will be used if  \code{method = "knn"}.}

\item{parallel}{Logical, if \code{TRUE} it use a parallel processing. \code{FALSE} as default.}
}
\value{
An object of class \code{neighborhood} which is a nested list that describe the cloud point used (\code{cloud.used}), the parameter \code{radius} or \code{k} used, and the resulting neighbor points per point (\code{neigborhood}).
}
\description{
Estimate neighboring points based on two methods.
}
\examples{
data("pc_tree")

###Estimate the niegborhood based in a sphere of a radius 0.2
neighborhood(pc_tree, method = "sphere", radius = 0.2, parallel = FALSE)

###Estimate the niegborhood based on knn
neighborhood(pc_tree, method = "knn", k = 10, parallel = FALSE)

###Parallel TRUE
require(doParallel)
detectCores() ### Number of cores of your computer
cores <- makeCluster(4) ### Set number of cores to work
registerDoParallel(cores)

neighborhood(pc_tree, method = "sphere", radius = 0.2, parallel = TRUE)

stopCluster(cores)

}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
