% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood.R
\name{neighborhood}
\alias{neighborhood}
\title{Neighboring points}
\usage{
neighborhood(cloud, cloud_b = NULL, method, radius, k,
  parallel = FALSE, threads = NULL)
}
\arguments{
\item{cloud}{A \code{data.table} with xyz coordinates in the first three columns.}

\item{cloud_b}{A \code{data.table} with xyz coordinates in the first three columns. If \code{cloud_b == NULL}, \code{cloud_b == cloud}. \code{NULL} as default.}

\item{method}{A character string specifying the method to estimated the neighbors. It most be one of  \code{"sphere"} or  \code{"knn"}.}

\item{radius}{An integer of a length 1 representing the number of neighbors to consider. This need to be used if  \code{method = "sphere"}, and this may usted if \code{method = "knn"}.}

\item{k}{An integer of a length 1 representing the number of neighbors to consider. This need to be used if \code{method = "knn"}.}

\item{parallel}{Logical, if \code{TRUE} it use a parallel processing. \code{FALSE} as default.}

\item{threads}{An \code{integer} >= 0 describing the number of CPU use. This need to be used if  \code{parallel = TRUE}}
}
\value{
An object of class \code{neighborhood} which is a nested list that describe the cloud point used (\code{cloud}), the parameter \code{radius} or \code{k} used, and the resulting neighbor points per point (\code{neigborhood}).
}
\description{
Estimate neighboring points based on two methods.
}
\examples{
data("pc_tree")

###Estimate the niegborhood based in a sphere of a radius 0.2
neighborhood(pc_tree[50:100,] , pc_tree, method = "sphere", radius = 0.2)

###Estimate the niegborhood based on knn
neighborhood(pc_tree, method = "knn", k = 10)

###Parallel TRUE with 4 threads
neighborhood(pc_tree, method = "sphere", radius = 0.2, parallel = TRUE, threads = 4)

}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
