% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood.R
\name{neighborhood}
\alias{neighborhood}
\title{Neighboring points}
\usage{
neighborhood(cloud, method, radius, k, parallel)
}
\arguments{
\item{cloud}{A  \code{matrix} or  \code{data.frame} with xyz coordinates in the first three columns.}

\item{method}{A character string specifying the method to estimated the neighbors. It most be one of  \code{"distance"} or  \code{"knn"}.}

\item{radius}{An integer of a length 1 representing the number of neighbors to consider. This will be used if  \code{method = "distance"}.}

\item{k}{An integer of a length 1 representing the number of neighbors to consider. This will be used if  \code{method = "knn"}.}

\item{parallel}{Logical, if  \code{TRUE} it use a parallel processing.}
}
\value{
An object of class \code{neighborhood} with a list of neighboring points per point.
}
\description{
Estimate neighboring points based on two methods.
}
\examples{
data("pc_tree")

###Estimate the niegborhood based on distance
neighborhood(pc_tree, method = "distance", radius = 0.2, parallel = FALSE)

###Estimate the niegborhood based on knn
neighborhood(pc_tree, method = "knn", k = 10, parallel = FALSE)

###Parallel TRUE
require(doParallel)
detectCores() ### Number of cores of your computer
cores <- makeCluster(4) ### Set number of cores to work
registerDoParallel(cores)

neighborhood(pc_tree, method = "distance", radius = 0.2, parallel = TRUE)

stopCluster(cores)

}
\references{
Wang, D., Hollaus, M., & Pfeifer, N. (2017). Feasibility of machine learning methods for separating wood and leaf points from Terrestrial Laser Scanning data. ISPRS Annals of Photogrammetry, Remote Sensing & Spatial Information Sciences, 4.
}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
