% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere_metrics.R
\name{sphere_metrics}
\alias{sphere_metrics}
\title{Sphere metrics on a target point}
\usage{
sphere_metrics(x, cloud, radius, basic = TRUE, distribution = TRUE,
  dimensionality = TRUE, n_replicates = NULL)
}
\arguments{
\item{x}{A \code{data.table} or \code{data.frame} of a targed point with xyz coordinates.}

\item{cloud}{A \code{data.table} of a point cloud with xyz coordinates to extract the neighboring points.}

\item{radius}{A numeric \code{vector} of a length 1 representing the minimum distance to consider neighboring points.}

\item{basic}{Logical, if \code{TRUE} it estimate basic metrics. \code{basic = TRUE} as default.}
}
\value{
A \code{data.table} with the xyz coordinates of the target point and the computed metrics by \code{basic.metrics()}, \code{distribution()}, and \code{dimensionality()} functions.
}
\description{
Estimate different metrics on a target point in a point cloud.
}
\examples{
data("pc_tree")
sphere_neighbors(pc_tree[100,], pc_tree, radius = 0.2)

}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
