% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_voxels.R
\name{summary_voxels}
\alias{summary_voxels}
\title{Voxels summary}
\usage{
summary_voxels(data, voxel.size, bootstrap, R)
}
\arguments{
\item{data}{A \code{data.frame} with xyz coordinates of the voxels in the first three columns and a fourth column with the number of points in each voxel.}

\item{voxel.size}{A positive \code{numeric} vector indicating the size used in \code{voxels}. It is needed to estimated metrics of density.1}

\item{bootstrap}{A logical vector length 1 vector. If \code{bootstrap = TRUE}, it compute a bootstrap on the H index calculations.}

\item{R}{A positive \code{integer} of length 1 indicating the number of bootstrap replicates.}
}
\value{
A \code{data.frame} with with the summary of \code{data}.
}
\description{
Create a summary of the voxels created using the function \code{voxels}.
}
\details{
The function estimate 9 main statistics of the voxels. Specifically, the first three columns (ei. \code(Voxel.size), \code(N_voxels), \code(Volumen)) describe the size of the voxel used, the number of voxels created, and the total volumen they represent.
Following columns represent the mean (\code{Density_mean}) and sd (\code{Density_sd}) of the density of points per voxel. Columns 6:10 provide metrics calculated using the Shannon index. Specifically, \code{H} describe the entropy, \code{H_max} the maximun entropy, \code{Equitavility} the ratio between \code{H} and \code{Hmax}, and \code{Negentropy} describe the product of \code{Hmax} - \code{H}.
If \code{bootstrap = TRUE} four more columns are created (10:13). These represent the mean and sd of the H index estimated using bootstrap (\code{H_boot_mean} and \code{H_boot_sd}), the \code{Equtavility_boot} as the ratio of the ratio between \code{H_boot_sd} and \code{Hmax}, and \code{Negentropy_boot} as the product \code{Hmax} - \code{H_boot_mean}.
}
\examples{
data("pc_tree")

###Create voxels of a size of 0.5.
vox <- voxels(pc_tree, voxel.size = 0.5)

###Summary voxels
summary_voxels(vox, voxel.size = 0.5, bootstrap = FALSE)

###Summary voxels and estimated the entropy using bootstrap with 1000 replicates
summary_voxels(vox, voxel.size = 0.5, bootstrap = TRUE, R = 1000)

}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
