% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLS_scan-data.R
\docType{data}
\name{TLS_scan}
\alias{TLS_scan}
\title{A TLS scan}
\format{A \code{data.table} with five columns, which are:
\describe{
\item{X}{the "X" coordinate}
\item{Y}{the "Y coordinate}
\item{Z}{the "Z" coordinate}
\item{Target_count}{The number of received by the same laser shot}
\item{Target_index}{The rank of the returned pulse in the target count of received by the same laser shot}
}}
\usage{
data(TLS_scan)
}
\description{
A \code{data.table} from a TLS scan.
}
\details{
A TLS scan conducted using a Reigel VZ400i with a vertical and horizonal resolution of 0.04 degrees. The scanner has frame of scanning between 30 and 130 degrees zenith and 0 and 360 degrees azimuth.
At the moment of the scan the roll, pitch, and yaw of the scanner were 0.293, -0.835, -150.159, respectivly. The scanner coodinates in this scan are \code{x = 0, y = 0, z = 0}.
}
\examples{
data(TLS_scan)
head(TLS_scan)

}
\keyword{datasets}
