% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxels.R
\name{voxels}
\alias{voxels}
\title{Voxelization of a Point Cloud}
\usage{
voxels(cloud, voxel.size, obj.voxels = TRUE)
}
\arguments{
\item{cloud}{A \code{matrix} or \code{data.frame} with *XYZ* coordinates in the first three columns.}

\item{voxel.size}{A positive \code{numeric} vector with the size of the voxel. It use the same dimentional scale of the point cloud.}

\item{obj.voxels}{Logical. If \code{obj.voxel = TRUE}, it returns an object of class \code{"voxels"}, If \code{obj.voxel = FALSE}, it returns a \code{data.table} with the coordinates of the voxels created and the number of points in each voxel. \code{TRUE} as default.}
}
\value{
If \code{TRUE}, it return an object of class \code{"voxels"} wich contain a list with the points used to create the voxels, the parameter \code{voxel.size}, and the \code{voxels} created. If \code{FALSE}, it returns a \code{data.table} with the coordinates of the voxels created and the number of points in each voxel.
}
\description{
Create cubes of a given distance in a point cloud though their voxelization. It use a modify version of the code used in Greaves et al. 2015.
}
\examples{
data("pc_tree")

###Create voxels of a size of 0.5.
voxels(pc_tree, voxel.size = 0.5)

###Returns just the coordinates of the voxels and the number of points in each voxel
voxels(pc_tree, voxel.size = 0.5, obj.voxel = FALSE)

}
\references{
Greaves, H. E., Vierling, L. A., Eitel, J. U., Boelman, N. T., Magney, T. S., Prager, C. M., & Griffin, K. L. (2015). Estimating aboveground biomass and leaf area of low-stature Arctic shrubs with terrestrial LiDAR. Remote Sensing of Environment, 164, 26-35.
}
\seealso{
\code{\link{voxels_counting}}, \code{\link{plot.voxels}}
}
\author{
J. Antonio Guzmán Q.
}
