% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxels_counting.R
\name{voxels_counting}
\alias{voxels_counting}
\title{Voxels counting}
\usage{
voxels_counting(cloud, voxel.range = NULL, random = FALSE,
  bootstrap = FALSE, R, parallel = FALSE, cores = NULL)
}
\arguments{
\item{cloud}{A \code{data.table} with xyz coordinates of the point clouds in the first three columns.}

\item{voxel.range}{A positive numeric vector describing the different voxel size to perform. If \code{voxel.range = NULL}, it use 10 voxel sizes by defaul. See details.}

\item{random}{Logical. If \code{TRUE}, it generates voxels on a set of random points created using the same number of points and _xyz_ range of \code{cloud}. \code{FALSE} as default.}

\item{bootstrap}{Logical, if \code{bootstrap = TRUE}, it computes a bootstrap on the H index calculations. \code{bootstrap = FALSE} as default.}

\item{R}{A positive \code{integer} of length 1 indicating the number of bootstrap replicates.}

\item{parallel}{Logical, if \code{TRUE} it uses a parallel processing. \code{FALSE} as default.}

\item{cores}{An \code{integer} >= 0 describing the number of cores to use. This need to be used if \code{parallel = TRUE}.}
}
\value{
A \code{data.table} with the summary of the voxels created with their features.
}
\description{
Creates voxels of different size on a point cloud using the \code{voxels()} function, and then return a summary of their features.
}
\examples{

data(pc_tree)

#Applying voxels counting
voxels_counting(pc_tree)

#Voxels counting using boostrap on the H indixes with 1000 repetitions
voxels_counting(pc_tree, bootstrap = TRUE, R = 1000)

#Voxels counting using bootstrap on the H indices with 1000 repetitions and applying parallel processing with 4 cores.
voxels_counting(pc_tree, bootstrap = TRUE, R = 1000, parallel = TRUE, cores = 4)

}
\author{
J. Antonio Guzman Q. and Ronny Hernandez
}
